+++ src/mapsManager.ts
@@ -1,10 +1,10 @@
 import * as vscode from 'vscode';
-import { ICompound, IDefined, fileToDefines, fileToNameToCompoundDefine, fileToNameToDefine } from './constants';
+import { GatherResults, ICompound, IDefined, fileToDefines, fileToNameToCompoundDefine, fileToNameToDefine } from './constants';
 import { gatherDefinitions } from "./parser";
 
-
-export async function updateFilesMapsIfEntries(document: vscode.TextDocument) {
-	const iDefineds: IDefined[] = await gatherDefinitions(document);
+export async function updateFilesMapsIfEntries(document:{doc?:vscode.TextDocument,uri?:vscode.Uri}) {
+	const gatherResults:GatherResults = (await gatherDefinitions(document?.doc ?? (await vscode.workspace.openTextDocument(document.uri))));
+	const iDefineds: IDefined[] = gatherResults.Defines
 	if (iDefineds.length) {
 		fileToDefines.set(document.uri, iDefineds)
 		let nameToCompound = new Map<string,ICompound>();
@@ -17,6 +17,6 @@ export async function updateFilesMapsIfEntries(document: vscode.TextDocument) {
 			}
 		}
 		if (nameToCompound.size) fileToNameToCompoundDefine.set(document.uri, nameToCompound)
-		if (nameToDefine.size) fileToNameToDefine.set(document.uri, nameToCompound)
+		if (nameToDefine.size) fileToNameToDefine.set(document.uri, nameToDefine)
 	}
 }